// crates/storyteller-workshop/src/lib/generated/HealthReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HealthStatus } from "./HealthStatus";
import type { SubsystemStatus } from "./SubsystemStatus";

/**
 * Server health report.
 */
export type HealthReport = { status: HealthStatus, subsystems: Array<SubsystemStatus>, };